[Php54] Rename ArrayToShortArrayRector to LongArrayToShortArrayRector
