+++ config/set/php54.php
@@ -3,13 +3,13 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php54\Rector\Array_\ArrayToShortArrayRector;
+use Rector\Php54\Rector\Array_\LongArrayToShortArrayRector;
 use Rector\Php54\Rector\Break_\RemoveZeroBreakContinueRector;
 use Rector\Php54\Rector\FuncCall\RemoveReferenceFromCallRector;
 use Rector\Renaming\Rector\FuncCall\RenameFunctionRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(ArrayToShortArrayRector::class);
+    $rectorConfig->rule(LongArrayToShortArrayRector::class);
 
     $rectorConfig
         ->ruleWithConfiguration(RenameFunctionRector::class, [
