--- rules-tests/Php54/Rector/Array_/ArrayToShortArrayRector/Fixture/skip_already_short_array.php.inc
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php54\Rector\Array_\ArrayToShortArrayTest\Fixture;
-
-function emptyArray()
-{
-    return [];
-}
-
-?>
