+++ rules-tests/Php54/Rector/Array_/LongArrayToShortArrayRector/Fixture/skip_already_short_array.php.inc
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rector\Tests\Php54\Rector\Array_\LongArrayToShortArrayTest\Fixture;
+
+function emptyArray()
+{
+    return [];
+}
+
+?>
