+++ rules-tests/Php54/Rector/Array_/LongArrayToShortArrayRector/config/configured_rule.php
@@ -3,8 +3,8 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php54\Rector\Array_\ArrayToShortArrayRector;
+use Rector\Php54\Rector\Array_\LongArrayToShortArrayRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(ArrayToShortArrayRector::class);
+    $rectorConfig->rule(LongArrayToShortArrayRector::class);
 };
