+++ rules/Php54/Rector/Array_/LongArrayToShortArrayRector.php
@@ -14,9 +14,9 @@ use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \Rector\Tests\Php54\Rector\Array_\ArrayToShortArrayRector\ArrayToShortArrayRectorTest
+ * @see \Rector\Tests\Php54\Rector\Array_\LongArrayToShortArrayRector\LongArrayToShortArrayRectorTest
  */
-final class ArrayToShortArrayRector extends AbstractRector implements MinPhpVersionInterface
+final class LongArrayToShortArrayRector extends AbstractRector implements MinPhpVersionInterface
 {
     public function provideMinPhpVersion(): int
     {
@@ -26,7 +26,7 @@ final class ArrayToShortArrayRector extends AbstractRector implements MinPhpVers
     public function getRuleDefinition(): RuleDefinition
     {
         return new RuleDefinition(
-            'Array to short array',
+            'Long array to short array',
             [
                 new CodeSample(
                     <<<'CODE_SAMPLE'
